+++ app/utils/types.py
@@ -5,8 +5,8 @@ from app.models import UserKarma, KarmaEvent, ModeratorEvent
 
 class ResultChangeKarma(NamedTuple):
     user_karma: UserKarma
-    abs_change: float
+    abs_change: int
     karma_event: KarmaEvent
     count_auto_restrict: int
-    karma_after: float
+    karma_after: int
     moderator_event: ModeratorEvent
